Started with MSAN and has a nice PIE implementation.
A working solution (missing documentation and perhaps some compile flag? And some test to veify it works).
I abondoned it since it requires lbcxx to be built with MSAN, but I could not get that to work, building it acc. to instructions (https://github.com/google/sanitizers/wiki/MemorySanitizerLibcxxHowTo) crashes with warnings from MSAN...
So I'll leave this here.

NOTE: MSAN requires code to be position independendent (compiled and linked with PIE-flag). This branch has a nice imeplementatoin of it (althpugh some compiler flag is set in execuatable_and_lib.cmake when is should be set in flags.cmake.